Rename script02_gen_newick_trees.py to generate_newick_trees.py
